# docs/swagger.yaml
definitions:
  carCombination.CarCombination:
    properties:
      ID:
        description: 主键ID
        type: integer
      childSeats:
        description: 儿童座椅数目
        type: integer
      combinationTitle:
        description: 车型组标题
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        type: integer
      deletedBy:
        type: integer
      largeLuggage:
        description: 24寸以上的箱包数量
        type: integer
      level:
        description: 车型级别
        type: string
      luggage:
        description: 24寸以下箱包数目
        type: integer
      modelName:
        description: 汽车车型
        type: string
      seats:
        description: 座位数量
        type: integer
      status:
        description: 车型组状态
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        type: integer
    required:
    - combinationTitle
    type: object
...
  order.Order:
    properties:
      ID:
        description: 主键ID
        type: integer
      appointment:
        description: 预约时间
        type: string
      cancelReason:
        description: 取消原因
        type: integer
      carModel:
        description: 车型组
        type: integer
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        type: integer
      deletedBy:
        type: integer
      fromArea:
        description: 开始地点
        type: string
      fromCity:
        description: 开始城市id
        type: integer
      orderSerial:
        description: 订单编号
        type: string
      passenger:
        description: 乘车人
        type: string
      passengerMobile:
        description: 乘车人联系
        type: string
      price:
        description: 订单金额
        type: integer
      status:
        description: 订单状态
        type: integer
      toArea:
        description: 抵达区域
        type: string
      toCity:
        description: 抵达城市ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        type: integer
    type: object
...
  response.Email:
    properties:
      body:
        description: 邮件内容
        type: string
      subject:
        description: 邮件标题
        type: string
      to:
        description: 邮件发送给谁
        type: string
    type: object
...
  response.OrderDetailResp:
    properties:
      car:
        $ref: '#/definitions/carCombination.CarCombination'
      order:
        $ref: '#/definitions/order.Order'
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.5.9
paths:
...
  /email/emailTest:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"发送成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 发送测试邮件
      tags:
      - System
  /email/sendEmail:
    post:
      parameters:
      - description: 发送邮件必须的参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/response.Email'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"发送成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 发送邮件
      tags:
      - System
...
  /web/car/detail:
    get:
      consumes:
      - application/json
      parameters:
      - description: 成人
        in: query
        name: aldult
        type: integer
      - description: 小孩
        in: query
        name: child
        type: integer
      - description: 一级父id
        in: query
        name: fromId
        type: integer
      - description: 一级父id
        in: query
        name: fromPid
        type: integer
      - in: query
        name: id
        type: integer
      - description: 行李
        in: query
        name: luggage
        type: integer
      - description: 出发时间
        in: query
        name: startTime
        type: string
      - in: query
        name: toId
        type: integer
      - in: query
        name: toPid
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      summary: 车型详情
      tags:
      - WebApi
  /web/car/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 成人
        in: query
        name: aldult
        type: integer
      - description: 小孩
        in: query
        name: child
        type: integer
      - in: query
        name: from
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 行李
        in: query
        name: luggage
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 出发时间
        in: query
        name: startTime
        type: string
      - in: query
        name: to
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      summary: 获取车刑列表
      tags:
      - WebApi
  /web/city/list:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: parentID
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      summary: 获取城市列表
      tags:
      - WebApi
...
  /web/order/cancel:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: cancelReason
        type: integer
      - in: query
        name: orderSerial
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"取消成功"}'
          schema:
            type: string
      summary: 取消订单
      tags:
      - WebApi
  /web/order/detail:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: orderSerial
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.OrderDetailResp'
        "400":
          description: '{"msg":"获取失败"}'
          schema:
            type: string
      summary: 订单详情
      tags:
      - WebApi
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"